Fix: Prevent undefined text length
The timeout within printChar to call the next instance was not passing the text along correctly. Originally recurring endlessly, an anonymous function was used to isolate the function call. Then the text was just passed in from the local scoped variable, instead of being passed in through the function arguments.